fix: Update README.md links (#2643)
## The basics

- [x] I [validated my changes](https://developers.google.com/blockly/guides/contribute/samples#making_and_verifying_a_change)

## The details
### Resolves

No issue filed for this.

### Proposed Changes

Updates the README to point to the correct new links

### Reason for Changes

While GitHub links automatically redirect, GitHub Pages do not and the old link may cause confusion for people who try using it.

### Test Coverage

I clicked on the link to make sure it works.

### Documentation

This is documentation.

### Additional Information

None.